refactor the relationship between workspace and packages
To use all of Atom's functionality, such as the git handling, we need to
add each package separately.

This commit modifies the plugin's structure to allow for just that. One
should add the overall workspace for "global" functionality - which is
right now the "Autoproj Update Workspace Info" command. Each package is
then added separately to get per-package functionality. The build
targets are filtered per-package (i.e. one gets only the targets
relevant for the package and/or for the overall workspace)